Voice/requirements.txt
openai-whisper
python-dotenv
sounddevice
scipy
pydub